Gross hack to set go build flag -buildvcs=false
Shopify build is failing with latest version of vitess which uses Go 1.18
when building docker images. This version of Go tries to embed version
control information when building inside a Git repository.
https://tip.golang.org/doc/go1.18#go-version

We previously added `.git` directories to the .dockerignore file to
work around an issue with buildx, so that's why this is failing for us,
but I'm not sure why it still thinks it's in a Git repository.

Can test this locally by running `make docker_vttestserver_mysql57`
which will fail with errors like:
```
error obtaining VCS status: exit status 128
	Use -buildvcs=false to disable VCS stamping.
```

Note that doesn't exercise buildx. Haven't tried that yet.

Also interesting to note there are some non-fatal errors like
```
fatal: not a git repository (or any of the parent directories): .git
```
These are likely caused by other scripts executing git commands as part of the vitess
build process, and I haven't looked into whether that's a problem.